# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/settings:
    get:
      operationId: ConfigController_getSettings
      summary: Retrieve settings
      description: Retrieve dynamic settings for this wallet
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeConfigDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: &a1
        - Settings
      security: &a2
        - oauth2:
            - wallet_issue_credentials
  /api/settings/update:
    post:
      operationId: ConfigController_updateSettings
      summary: Update settings
      description: Update the dynamic settings for this wallet
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuntimeConfigDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeConfigDto"
        "400":
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a1
      security: *a2
...
  /context/{id}:
    get:
      operationId: ContextController_getContext
      summary: Retrieve context
      description: Retrieves JSON-LD context document for the given context
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags:
        - Contexts
  /api/management/contexts:
    get:
      operationId: ContextManagementController_getContexts
      summary: Retrieve contexts
      description: Retrieve all context registered in this wallet
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JSONLDContextDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: &a16
        - Management Contexts
      security: &a17
        - oauth2:
            - wallet_view_did
    post:
      operationId: ContextManagementController_addContext
      summary: Add context
      description: Register a new context in this wallet
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonLdContextConfigDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JSONLDContextDto"
        "400":
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "409":
          description: Resource already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a16
      security: *a17
  /api/management/contexts/{id}:
    put:
      operationId: ContextManagementController_updateContext
      summary: Update context
      description: Update an existing context in this wallet
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JsonLdContextConfigDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JSONLDContextDto"
        "400":
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a16
      security: *a17
    delete:
      operationId: ContextManagementController_deleteContext
      summary: Delete context
      description: Delete an existing context in this wallet
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
      tags: *a16
      security: *a17
...
info:
  title: TSG Wallet
  description: Description
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        password:
          scopes:
            wallet_view_did: wallet_view_did
            wallet_manage_keys: wallet_manage_keys
            wallet_view_own_credentials: wallet_view_own_credentials
            wallet_view_all_credentials: wallet_view_all_credentials
            wallet_manage_own_credentials: wallet_manage_own_credentials
            wallet_manage_all_credentials: wallet_manage_all_credentials
            wallet_issue_credentials: wallet_issue_credentials
            wallet_view_presentations: wallet_view_presentations
  schemas:
    RuntimeConfigDto:
      type: object
      properties:
        gaiaXSupport:
          type: boolean
      required:
        - gaiaXSupport
    ErrorDto:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        code:
          type: number
        message:
          type: object
        error:
          type: string
      required:
        - name
        - status
        - code
...
    JSONLDContextDto:
      type: object
      properties:
        id:
          type: string
        credentialType:
          type: string
        issuable:
          type: boolean
        documentUrl:
          type: string
        document:
          type: object
        schema:
          type: object
        created:
          format: date-time
          type: string
        modified:
          format: date-time
          type: string
        deleted:
          format: date-time
          type: string
      required:
        - id
        - credentialType
        - issuable
        - created
        - modified
    JsonLdContextConfigDto:
      type: object
      properties:
        id:
          type: string
        credentialType:
          type: string
        issuable:
          type: boolean
        documentUrl:
          type: string
        document:
          type: object
        schema:
          type: object
      required:
        - id
        - credentialType
        - issuable
...